+++ inc/MathFunctions.hpp
@@ -83,8 +83,8 @@ Eigen::Vector3f cartesianToSpherical(Eigen::Vector3f vectorCartesian);
  */
 Eigen::Vector3f sphericalToCartesian(Eigen::Vector3f vectorSpherical);
 
-Eigen::Vector3d eci_to_ecef(Eigen::Vector3d vec_eci, double gstime);
+Eigen::Vector3f eci_to_ecef(Eigen::Vector3f vec_eci, double gstime);
 
-Eigen::Vector3d ecef_to_llh(Eigen::Vector3d uvw);
+Eigen::Vector3f ecef_to_llh(Eigen::Vector3f uvw);
